extract pair-wise pattern from ib_validation
