+++ docs/superbench-config.mdx
@@ -463,4 +463,4 @@ Only available for `mpi` mode.
 
 Available variables in formatted string includes:
 + `name`
-    * accepted values: `all-nodes`
+    * accepted values: `all-nodes`, `pair-wise`
